# docs/Swagger.yaml
swagger: '2.0'
info:
  description: |
    AND1 Project
  version: pre-alpha-1.0
  title: Event & Equipment Manager
paths:
...
  /equipment:
    get:
      summary: returns all equipments
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - in: query
        name: eventId
        type: integer
      produces:
      - application/json
      responses:
        200:
          description: Array of Equipments
          schema:
            type: array
            items:
              $ref: '#/definitions/equipment'
        401:
          description: Unauthorized
    post:
      summary: creates new equipment
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - in: body
        name: Equipment
        required: true
        description: Equipment
        schema:
          $ref: '#/definitions/equipment'
      produces:
      - text/plain
      responses:
        200:
          description: Created
        401:
          description: Unauthorized
  /equipment/{EquipmentId}:
    get:
      summary: returns equipment with given id
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: EquipmentId
        in: path
        description: Id of the equipment to return
        required: true
        type: integer
      produces:
      - application/json
      responses:
        200:
          description: Equipment
          schema:
            $ref: '#/definitions/equipment'
        401:
          description: Unauthorized
        404:
          description: Not Found
    delete:
      summary: deletes equipment with given id
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: EquipmentId
        in: path
        description: Id of the equipment to delete
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        200:
          description: Deleted
        401:
          description: Unauthorized
        404:
          description: Not Found
    put:
      summary: updates equipment with given id
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: EquipmentId
        in: path
        description: Id of the equipment to update
        required: true
        type: integer
      - in: body
        name: Event
        required: true
        description: Equipment with updated attributes
        schema:
          $ref: '#/definitions/equipment'
      produces:
      - text/plain
      responses:
        200:
          description: Updated
        401:
          description: Unauthorized
        404:
          description: Not Found
        409:
          description: "Conflict, descriptive error message detailing which events are conflicting in the involved equipment's scope."
          schema:
            example: "Event 0 (id) \"Generali Night Run Budapest\" conflicts with Event 1 (id) \"Sólet Fesztivál\"\nEvent 2 (id) \"Kolbi Fesztivál\" conflicts with Event 3 (id) \"Bableves Fesztivál\""
definitions:
...
  equipment:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: Yamaha LS9-32
      category:
        type: string
        example: Audio Racks
      events:
        type: array
        items:
          type: object
          properties:
            id:
              type: integer
              example: 1
            name:
              type: string
              example: Generali Night Run Budapest
            hidden:
              type: boolean
              example: false
              description: There are some autogenerated events like repair periods for equipment, which are only manageable indirectly. These events only show up in the equipment’s reserved history, while being hidden in the events calendar.
            start:
              type: integer
              example: 1492992000000
              description: "DateTime, format: number of milliseconds since the standard base time January 1, 1970, 00:00:00 GMT."
            end:
              type: integer
              example: 1493078400000
              description: "DateTime, format: number of milliseconds since the standard base time January 1, 1970, 00:00:00 GMT."
...